+++ index/common_test.go
@@ -59,7 +59,7 @@ func testingAPIServer(
 
 	reg := prometheus.NewRegistry()
 
-	instrumentation, err := elephantine.NewHTTPClientIntrumentation(reg)
+	instrumentation, err := elephantine.NewHTTPClientInstrumentation(reg)
 	test.Mustf(t, err, "set up HTTP client instrumentation")
 
 	env := SetUpBackingServices(t, instrumentation, false)